backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "pandas>=2.3.1",
    "pymongo-stubs>=0.2.0",
    "pymongo[srv]==3.12",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pandas-stubs>=2.3.0.250703",
    "scipy-stubs>=1.16.0.2",
    "types-flask-cors>=6.0.0.20250520",
    "types-openpyxl>=3.1.5.20250602",
    "types-python-dateutil>=2.9.0.20250708",
    "types-requests>=2.32.4.20250611",
]
